+++ src/main/java/atdd/station/model/CreateStationRequestView.java
@@ -1,0 +1,13 @@
+package atdd.station.model;
+
+import lombok.Setter;
+
+@Setter
+public class CreateStationRequestView {
+    private String name;
+
+    public Station toStation() {
+        return Station.builder()
+                .name(this.name).build();
+    }
+}
